+++ src/actions.py
@@ -55,25 +55,6 @@ ctx.lists["self.cursorless_simple_action"] = {
     # "paste": "paste",
 }
 
-wrappers = {
-    "square": ["[", "]"],
-    "round": ["(", ")"],
-    "curly": ["{", "}"],
-    "angle": ["<", ">"],
-    "twin": ["'", "'"],
-    "quad": ['"', '"'],
-    "brick": ["`", "`"],
-    "escaped twin": ["\\'", "\\'"],
-    "escaped quad": ['\\"', '\\"']
-}
-
-mod.list("cursorless_wrapper", desc="Supported wrappers for cursorless wrap action")
-ctx.lists["self.cursorless_wrapper"] = wrappers.keys()
-
-@mod.capture(rule=("{user.cursorless_wrapper}"))
-def cursorless_wrapper(m) -> list[str]:
-    return wrappers[m.cursorless_wrapper]
-
 
 @mod.capture(rule=("swap [<user.cursorless_target>] with <user.cursorless_target>"))
 def cursorless_swap(m) -> str:
